AI/conversation/tutor.py
import os
from typing import Dict, List

class LanguageTutor:
    def __init__(self, personality: str, teaching_style: str, language: str):
        self.personality = personality
        self.teaching_style = teaching_style
        self.language = language
        self.conversation_history: List[Dict] = []

    def generate_response(self, user_input: str) -> str:
        """Generate a response based on user input and tutor's personality."""
        # Placeholder for actual GPT integration
        prompt = self._create_prompt(user_input)
        response = self._get_model_response(prompt)
        self.conversation_history.append({
            'user': user_input,
            'tutor': response
        })
        return response

    def _create_prompt(self, user_input: str) -> str:
        """Create a prompt for the language model based on context."""
        context = f"You are a {self.personality} language tutor teaching {self.language}. "
        context += f"Your teaching style is {self.teaching_style}. "
        return context + user_input

    def _get_model_response(self, prompt: str) -> str:
        """Get response from language model."""
        # Placeholder for actual API call
        return f"This is a placeholder response for: {prompt}"

    def evaluate_response(self, user_response: str) -> Dict:
        """Evaluate user's language use and provide feedback."""
        return {
            'accuracy': 0.8,  # Placeholder
            'grammar': 0.9,
            'vocabulary': 0.7,
            'feedback': 'Good job! Here are some suggestions...'
        }